Validate and update vote reservation timestamp
This is done after vote generation to be correct, and the generated votes are dropped if validation failed (should not normally happen, but is in place to ensure correctness).

This fixes intermittent failures on `vote_generator.spacing` which came from the stamps on the generated votes being less than **round_time** apart.